refactor: Rename to ManualFromScaleOrFromOffsetLint
