+++ selene-lib/src/lints.rs
@@ -37,7 +37,7 @@ pub mod unscoped_variables;
 pub mod unused_variable;
 
 #[cfg(feature = "roblox")]
-pub mod roblox_complex_udim2_new;
+pub mod roblox_manual_fromscale_or_fromoffset;
 
 #[cfg(feature = "roblox")]
 pub mod roblox_incorrect_color3_new_bounds;
